+++ src/components/logicgates/LogicGates.java
@@ -42,4 +42,13 @@ public interface LogicGates extends LogicGatesKernel {
      *          newInput))
      */
     LogicGates replaceInput(LogicGates newInput, boolean leftGate);
+
+    /**
+     * Computes the final boolean value for the entirety of {@code this}, and
+     * returns said boolean.
+     *
+     * @return the final boolean value
+     * @ensures compute = compute(first) operator compute(second)
+     */
+    boolean compute();
 }